examples/python-sql/requirements.txt
mysql-connector-python
qrcode[pil]
pyqrcode
